add Burial blade descriptions
- added item description:
A strange, yet beautiful, hybrid of sword and scythe.

Discovered in an ancient tomb beneath a Carim cathedral, some say it's magnificent design inspired the curved swords Carim knights are known for.
But, such accolades may simply be the words of an unscrupulous merchant hoping to sell it to a naïve customer.

Skill: Spinning Motion/Neck Swipe
Execute a deliberate slashing move. When transformed, attack towards the scruff of a foe's neck, and when successful, functions as a head shot, inflicting heavy damage.

- changed weapon art in the menu to Spinning Motion / Neck Swipe