Updated burial blade descriptions
A strange, yet beautiful, hybrid of sword and scythe.

Discovered in an ancient tomb beneath a Carim cathedral, some say it's magnificent design inspired the curved swords Carim knights are known for.
But, such accolades may simply be the words of an unscrupulous merchant hoping to sell it to an unsuspecting customer.

Skill: Spinning Motion/Reaping Slashes
Execute a deliberate horizontal attack. When transformed, perform two powerful slashes in rapid succession.