add amygdalan arm description
A uniquely grotesque and brutal instrument.

The Amygdalan Arm earned it's name from the madmen who wielded it, drawing etymological inspiration from the amygdala, an area of the brain associated with primitive emotions, and its unsavoury severance from the carcass of some horrible beast best forgotten.

It can be wielded as a blunt club of sorts, but when extended the beast within stirs, quivering as if yet still living.

Skill: Nightmare Serpent
Thrust the weapon before you, drawing on some terrible lingering power to let forth a shockwave of energy.