Edit Dragon Bone Smasher Attacks to work with poise resisted throws
all throws that should be poise resisted
(everything except charged R2 and WA)
Now use 0x18B (pvpstun) 4 (ss) and have behavior2 set to 26 or 28 for throw/panckake respectively.

The ss stun will never be used, it's just there for the poise check. If it passes the poise check it will be converted to a throw.