Added description for ironclad #967
An old armor that is oddly comfortable. Provides high defense, but is so heavy that it requires great strength to move.

Feint traces of an ancient enchantment remain in the shell. Faded by time, it is again an inert iron husk, fit to wear.

(only on the chest piece)
Old as it may be, the iron shell of this armor still protects its wearers back.