Crown of Filianore rewrite
Crown of Filianore, youngest daughter of Gwyn, Lord of Cinder.
The Spears of the Church as her guardians, her slumber was left uninterrupted.

There is no deeper agony for a fool than a stone left unturned; the pure absence of closure.
There is no greater comedy than the same fool at his end, in a world entombed by colored sands.